WA-5145 open read only documents checked out by others
The CAN_SET_CONTENT_STREAM is allowed in Alfresco on latest
version(which is the PWC in this case) but it is not allowed in
SharePoint (only the user who checked out the document can 'see' the
PWC). In SharePoint the login user name does not match the
cmis:versionSeriesCheckedOutBy attribute. Fixed to open in read only
mode for both cases.